\name{cor.pba}
\alias{cor.pba}
\title{Correlation method for objects of class pba...}
\usage{cor.pba(pba, ...)cor.pba(pba, ...)
cor.pba(pba, ...)
cor.pba(pba, ...)
cor.pba(pba, ...)
cor.pba(pba, ...)
cor.pba(pba, ...)
cor.pba(pba, ...)
cor.pba(pba, ...)
cor.pba(pba, ...)
}
\description{Correlation method for objects of class \code{pba}}
\value{Returns a list of correlation matrices of the defined biases.\code{cor.pba}: Returns a list of correlation matrices of the defined biases.


\code{cor.pba}: Returns a list of correlation matrices of the defined biases.


\code{cor.pba}: Returns a list of correlation matrices of the defined biases.


\code{cor.pba}: Returns a list of correlation matrices of the defined biases.


\code{cor.pba}: Returns a list of correlation matrices of the defined biases.


\code{cor.pba}: Returns a list of correlation matrices of the defined biases.


\code{cor.pba}: Returns a list of correlation matrices of the defined biases.


\code{cor.pba}: Returns a list of correlation matrices of the defined biases.


\code{cor.pba}: Returns a list of correlation matrices of the defined biases.


}
\arguments{\item{An}{object of class \code{pba}}
\item{...}{Further arguments to pass to \code{cor()}}}
\alias{cor.pba}

\alias{cor.pba}

\alias{cor.pba}

\alias{cor.pba}

\alias{cor.pba}

\alias{cor.pba}

\alias{cor.pba}

\alias{cor.pba}

\alias{cor.pba}

