\name{parse.ref.preref}
\alias{parse.ref.preref}
\alias{parse.ref.srcref}
\title{Parse a preref...}
\usage{\method{parse.ref}{preref} (ref, ...)
\method{parse.ref}{srcref} (...)
}
\description{Parse a preref}
\value{List containing the parsed preref}
\arguments{\item{ref}{the preref to be parsed}
\item{\dots}{ignored}}
