\name{pbaDistr}
\alias{pbaDistr}
\title{Function to define distributions for bias parameters.}
\usage{pbaDistr(distr, ...)pbaDistr(distr, ...)
pbaDistr(distr, ...)
pbaDistr(distr, ...)
pbaDistr(distr, ...)
pbaDistr(distr, ...)
pbaDistr(distr, ...)
pbaDistr(distr, ...)
pbaDistr(distr, ...)
pbaDistr(distr, ...)
}
\description{Function to define distributions for bias parameters.}
\value{An object of class \code{pba.distr}.\code{pbaDistr}: An object of class \code{pba.distr}.


\code{pbaDistr}: An object of class \code{pba.distr}.


\code{pbaDistr}: An object of class \code{pba.distr}.


\code{pbaDistr}: An object of class \code{pba.distr}.


\code{pbaDistr}: An object of class \code{pba.distr}.


\code{pbaDistr}: An object of class \code{pba.distr}.


\code{pbaDistr}: An object of class \code{pba.distr}.


\code{pbaDistr}: An object of class \code{pba.distr}.


\code{pbaDistr}: An object of class \code{pba.distr}.


}
\author{Jeremy Thoms Hetzel \email{jthetzel@gmail.com}
{\item{distr}{A character vector of length one naming a distribution function.
For example `norm' for the normal distribution.}
\item{...}{Additional arguments to be passed to the \code{distr} function.}}
}
\alias{pbaDistr}
\arguments{\item{distr}{A character vector of length one naming a distribution function.
For example `norm' for the normal distribution.}
\item{...}{Additional arguments to be passed to the \code{distr} function.}
}

\alias{pbaDistr}

\alias{pbaDistr}

\alias{pbaDistr}

\alias{pbaDistr}

\alias{pbaDistr}

\alias{pbaDistr}

\alias{pbaDistr}

\alias{pbaDistr}

