\name{pbaPlotConfoundingRisks}
\alias{pbaPlotConfoundingRisks}
\title{Plot density distribution of confounding risk bias parameters of a pba object...}
\usage{pbaPlotConfoundingRisks(pba, data, parameters=c("rr\", \"rd"),
    density=T, title="Confounding (relative risks)", scales="free",
    L2="confounding")pbaPlotConfoundingRisks(pba, data, parameters=c("rr\", \"rd"), density=T, title="Confounding
    (relative risks)", scales="free", L2="confounding")
pbaPlotConfoundingRisks(pba, data, parameters=c("rr\", \"rd"), density=T, title="Confounding
    (relative risks)", scales="free", L2="confounding")
pbaPlotConfoundingRisks(pba, data, parameters=c("rr\", \"rd"), density=T, title="Confounding
    (relative risks)", scales="free", L2="confounding")
pbaPlotConfoundingRisks(pba, data, parameters=c("rr\", \"rd"), density=T, title="Confounding
    (relative risks)", scales="free", L2="confounding")
pbaPlotConfoundingRisks(pba, data, parameters=c("rr\", \"rd"), density=T, title="Confounding
    (relative risks)", scales="free", L2="confounding")
pbaPlotConfoundingRisks(pba, data, parameters=c("rr\", \"rd"), density=T, title="Confounding
    (relative risks)", scales="free", L2="confounding")
pbaPlotConfoundingRisks(pba, data, parameters=c("rr\", \"rd"), density=T, title="Confounding
    (relative risks)", scales="free", L2="confounding")
pbaPlotConfoundingRisks(pba, data, parameters=c("rr\", \"rd"), density=T, title="Confounding
    (relative risks)", scales="free", L2="confounding")
pbaPlotConfoundingRisks(pba, data, parameters=c("rr\", \"rd"), density=T, title="Confounding
    (relative risks)", scales="free", L2="confounding")
}
\description{Plot density distribution of confounding risk bias parameters of a pba object}
\value{The \code{plotBiasConfoundingRisks} function returns density plots or histograms of the
distribution of confounding relative risk or risk difference bias parameters for a \code{pba} object. 
This function allows finer control of plots than the \code{plotBias} function.\code{pbaPlotConfoundingRisks}: The \code{plotBiasConfoundingRisks} function returns density plots or histograms of the
distribution of confounding relative risk or risk difference bias parameters for a \code{pba} object. 
This function allows finer control of plots than the \code{plotBias} function.


\code{pbaPlotConfoundingRisks}: The \code{plotBiasConfoundingRisks} function returns density plots or histograms of the
distribution of confounding relative risk or risk difference bias parameters for a \code{pba} object. 
This function allows finer control of plots than the \code{plotBias} function.


\code{pbaPlotConfoundingRisks}: The \code{plotBiasConfoundingRisks} function returns density plots or histograms of the
distribution of confounding relative risk or risk difference bias parameters for a \code{pba} object. 
This function allows finer control of plots than the \code{plotBias} function.


\code{pbaPlotConfoundingRisks}: The \code{plotBiasConfoundingRisks} function returns density plots or histograms of the
distribution of confounding relative risk or risk difference bias parameters for a \code{pba} object. 
This function allows finer control of plots than the \code{plotBias} function.


\code{pbaPlotConfoundingRisks}: The \code{plotBiasConfoundingRisks} function returns density plots or histograms of the
distribution of confounding relative risk or risk difference bias parameters for a \code{pba} object. 
This function allows finer control of plots than the \code{plotBias} function.


\code{pbaPlotConfoundingRisks}: The \code{plotBiasConfoundingRisks} function returns density plots or histograms of the
distribution of confounding relative risk or risk difference bias parameters for a \code{pba} object. 
This function allows finer control of plots than the \code{plotBias} function.


\code{pbaPlotConfoundingRisks}: The \code{plotBiasConfoundingRisks} function returns density plots or histograms of the
distribution of confounding relative risk or risk difference bias parameters for a \code{pba} object. 
This function allows finer control of plots than the \code{plotBias} function.


\code{pbaPlotConfoundingRisks}: The \code{plotBiasConfoundingRisks} function returns density plots or histograms of the
distribution of confounding relative risk or risk difference bias parameters for a \code{pba} object. 
This function allows finer control of plots than the \code{plotBias} function.


\code{pbaPlotConfoundingRisks}: The \code{plotBiasConfoundingRisks} function returns density plots or histograms of the
distribution of confounding relative risk or risk difference bias parameters for a \code{pba} object. 
This function allows finer control of plots than the \code{plotBias} function.


}
\author{Jeremy Thoms Hetzel \email{jthetzel@gmail.com}}
\arguments{\item{pba}{An object of class \code{pba}.}
\item{data}{(optional) If \code{pba} is not provided, a data frame object can
be specified. The optional argument is provided ofr users desiring finer 
control of the plots.}
\item{density}{Logical; if true, plots smoothed density. If false, plots
histogram.}
\item{parameters}{A character vector specifying which bias parameters to
include in the plots. Expected values are `rr' and `rd'.}
\item{title}{A character vector of length one specifying the plot's title.}
\item{sclaes}{Character string passed to ggplot. This parameter controls
whether the x and y axes are free or fixed. Acceptable values are "free",
"fixed", "free_x", or "free_y"}
\item{L2}{A character vector of length one specifying which bias type to plot.
The expected value for this function is `confounding'.}}
\alias{pbaPlotConfoundingRisks}

\alias{pbaPlotConfoundingRisks}

\alias{pbaPlotConfoundingRisks}

\alias{pbaPlotConfoundingRisks}

\alias{pbaPlotConfoundingRisks}

\alias{pbaPlotConfoundingRisks}

\alias{pbaPlotConfoundingRisks}

\alias{pbaPlotConfoundingRisks}

\alias{pbaPlotConfoundingRisks}

