\name{pbaPlotSelection}
\alias{pbaPlotSelection}
\title{Plot density distribution of selection bias parameters of a pba object...}
\usage{pbaPlotSelection(pba, data, density=T, parameters=c("s.a1\", \"s.a0\",
    \"s.b1\", \"s.b0"), title="Selection", scales="free",
    L2="selection")pbaPlotSelection(pba, data, density=T, parameters=c("s.a1\", \"s.a0\", \"s.b1\",
    \"s.b0"), title="Selection", scales="free", L2="selection")
pbaPlotSelection(pba, data, density=T, parameters=c("s.a1\", \"s.a0\", \"s.b1\",
    \"s.b0"), title="Selection", scales="free", L2="selection")
pbaPlotSelection(pba, data, density=T, parameters=c("s.a1\", \"s.a0\", \"s.b1\",
    \"s.b0"), title="Selection", scales="free", L2="selection")
pbaPlotSelection(pba, data, density=T, parameters=c("s.a1\", \"s.a0\", \"s.b1\",
    \"s.b0"), title="Selection", scales="free", L2="selection")
pbaPlotSelection(pba, data, density=T, parameters=c("s.a1\", \"s.a0\", \"s.b1\",
    \"s.b0"), title="Selection", scales="free", L2="selection")
pbaPlotSelection(pba, data, density=T, parameters=c("s.a1\", \"s.a0\", \"s.b1\",
    \"s.b0"), title="Selection", scales="free", L2="selection")
pbaPlotSelection(pba, data, density=T, parameters=c("s.a1\", \"s.a0\", \"s.b1\",
    \"s.b0"), title="Selection", scales="free", L2="selection")
pbaPlotSelection(pba, data, density=T, parameters=c("s.a1\", \"s.a0\", \"s.b1\",
    \"s.b0"), title="Selection", scales="free", L2="selection")
pbaPlotSelection(pba, data, density=T, parameters=c("s.a1\", \"s.a0\", \"s.b1\",
    \"s.b0"), title="Selection", scales="free", L2="selection")
}
\description{Plot density distribution of selection bias parameters of a pba object}
\value{The \code{plotBiasSelection} function returns density plots or histograms of the
distribution of selection bias parameters for a \code{pba} object. 
This function allows finer control of plots than the \code{plotBias} function.\code{pbaPlotSelection}: The \code{plotBiasSelection} function returns density plots or histograms of the
distribution of selection bias parameters for a \code{pba} object. 
This function allows finer control of plots than the \code{plotBias} function.


\code{pbaPlotSelection}: The \code{plotBiasSelection} function returns density plots or histograms of the
distribution of selection bias parameters for a \code{pba} object. 
This function allows finer control of plots than the \code{plotBias} function.


\code{pbaPlotSelection}: The \code{plotBiasSelection} function returns density plots or histograms of the
distribution of selection bias parameters for a \code{pba} object. 
This function allows finer control of plots than the \code{plotBias} function.


\code{pbaPlotSelection}: The \code{plotBiasSelection} function returns density plots or histograms of the
distribution of selection bias parameters for a \code{pba} object. 
This function allows finer control of plots than the \code{plotBias} function.


\code{pbaPlotSelection}: The \code{plotBiasSelection} function returns density plots or histograms of the
distribution of selection bias parameters for a \code{pba} object. 
This function allows finer control of plots than the \code{plotBias} function.


\code{pbaPlotSelection}: The \code{plotBiasSelection} function returns density plots or histograms of the
distribution of selection bias parameters for a \code{pba} object. 
This function allows finer control of plots than the \code{plotBias} function.


\code{pbaPlotSelection}: The \code{plotBiasSelection} function returns density plots or histograms of the
distribution of selection bias parameters for a \code{pba} object. 
This function allows finer control of plots than the \code{plotBias} function.


\code{pbaPlotSelection}: The \code{plotBiasSelection} function returns density plots or histograms of the
distribution of selection bias parameters for a \code{pba} object. 
This function allows finer control of plots than the \code{plotBias} function.


\code{pbaPlotSelection}: The \code{plotBiasSelection} function returns density plots or histograms of the
distribution of selection bias parameters for a \code{pba} object. 
This function allows finer control of plots than the \code{plotBias} function.


}
\author{Jeremy Thoms Hetzel \email{jthetzel@gmail.com}}
\arguments{\item{pba}{An object of class \code{pba}.}
\item{data}{(optional) If \code{pba} is not provided, a data frame object can
be specified. The optional argument is provided ofr users desiring finer 
control of the plots.}
\item{density}{Logical; if true, plots smoothed density. If false, plots
histogram.}
\item{parameters}{A character vector specifying which bias parameters to
include in the plots. Expected values are `s.a1', `s.b1', `s.a0', and `s.b0'.}
\item{title}{A character vector of length one specifying the plot's title.}
\item{sclaes}{Character string passed to ggplot. This parameter controls
whether the x and y axes are free or fixed. Acceptable values are "free",
"fixed", "free_x", or "free_y"}
\item{L2}{A character vector of length one specifying which bias type to plot.
The expected value for this function is `selection'.}}
\alias{pbaPlotSelection}

\alias{pbaPlotSelection}

\alias{pbaPlotSelection}

\alias{pbaPlotSelection}

\alias{pbaPlotSelection}

\alias{pbaPlotSelection}

\alias{pbaPlotSelection}

\alias{pbaPlotSelection}

\alias{pbaPlotSelection}

