\name{plot.pba}
\title{Plot density distribution of estimates after adjusting for bias...}
\usage{plot.pba(pba, data, density=T, exp=F, adjust=1, binwidth,
    scales="free", variables, print=T)plot.pba(pba, data, density=T, exp=F, adjust=1, binwidth, scales="free",
    variables, print=T)
plot.pba(pba, data, density=T, exp=F, adjust=1, binwidth, scales="free",
    variables, print=T)
plot.pba(pba, data, density=T, exp=F, adjust=1, binwidth, scales="free",
    variables, print=T)
plot.pba(pba, data, density=T, exp=F, adjust=1, binwidth, scales="free",
    variables, print=T)
plot.pba(pba, data, density=T, exp=F, adjust=1, binwidth, scales="free",
    variables, print=T)
plot.pba(pba, data, density=T, exp=F, adjust=1, binwidth, scales="free",
    variables, print=T)
}
\description{Plot density distribution of estimates after adjusting for bias}
\value{The plotEstimates function returns density plots or histograms of the
coefficient estimates.\code{plot.pba}: The plotEstimates function returns density plots or histograms of the
coefficient estimates.


\code{plot.pba}: The plotEstimates function returns density plots or histograms of the
coefficient estimates.


\code{plot.pba}: The plotEstimates function returns density plots or histograms of the
coefficient estimates.


\code{plot.pba}: The plotEstimates function returns density plots or histograms of the
coefficient estimates.


\code{plot.pba}: The plotEstimates function returns density plots or histograms of the
coefficient estimates.


\code{plot.pba}: The plotEstimates function returns density plots or histograms of the
coefficient estimates.


}
\alias{plotEstimates}
\author{Jeremy Thoms Hetzel \email{jthetzel@gmail.com}}
\arguments{\item{pba}{A pba object.}
\item{data}{(optional) If \code{pba} is not provided, an appropriate data frame
may be provided here instead. This option is intended for users requiring 
finer control.
#'}
\item{density}{Logical; if true, plots smoothed density. If false, plots
histogram.}
\item{exp}{Logical; if true, estimates are exponentiated.}
\item{adjust}{Passed to \code{density()}}
\item{binwidth}{Passed to \code{histogram()}}
\item{sclaes}{Character string passed to ggplot. This parameter controls
whether the x and y axes are free or fixed. Acceptable values are "free",
"fixed", "free_x", or "free_y"}
\item{Variables}{A character vector specifying the variables to plot. If
NULL, the function will plot all variables.}
\item{print}{Logical; if true, function prints the plots. If false, the
function only returns the plot objects in a list.}}
\alias{plot.pba}
\alias{plotEstimates}

\alias{plot.pba}
\alias{plotEstimates}

\alias{plot.pba}

\alias{plot.pba}

\alias{plot.pba}

\alias{plot.pba}

