\name{plotBias}
\alias{plotBias}
\title{Plot density distribution of bias parameters of a pba object...}
\usage{plotBias(pba, density=T, scales="free", types, print=T)plotBias(pba, density=T, scales="free", types, print=T)
plotBias(pba, density=T, scales="free", types, print=T)
plotBias(pba, density=T, scales="free", types, print=T)
plotBias(pba, density=T, scales="free", types, print=T)
plotBias(pba, density=T, scales="free", types, print=T)
plotBias(pba, density=T, scales="free", types, print=T)
plotBias(pba, density=T, scales="free", types, print=T)
plotBias(pba, density=T, scales="free", types, print=T)
plotBias(pba, density=T, scales="free", types, print=T)
}
\description{Plot density distribution of bias parameters of a pba object}
\value{The plotBias function returns density plots or histograms of the
distribution of bias parameters for a pba object. If print = TRUE, the
function attempts to organize all plots into a single window. Individual 
plots may be accessed from the returned list. plotBias internally calls
pbaPlotMisclassification, pbaPlotSelection, pbaPlotConfoundingProportions,
and pbaPlotConfoundingRisks.\code{plotBias}: The plotBias function returns density plots or histograms of the
distribution of bias parameters for a pba object. If print = TRUE, the
function attempts to organize all plots into a single window. Individual 
plots may be accessed from the returned list. plotBias internally calls
pbaPlotMisclassification, pbaPlotSelection, pbaPlotConfoundingProportions,
and pbaPlotConfoundingRisks.


\code{plotBias}: The plotBias function returns density plots or histograms of the
distribution of bias parameters for a pba object. If print = TRUE, the
function attempts to organize all plots into a single window. Individual 
plots may be accessed from the returned list. plotBias internally calls
pbaPlotMisclassification, pbaPlotSelection, pbaPlotConfoundingProportions,
and pbaPlotConfoundingRisks.


\code{plotBias}: The plotBias function returns density plots or histograms of the
distribution of bias parameters for a pba object. If print = TRUE, the
function attempts to organize all plots into a single window. Individual 
plots may be accessed from the returned list. plotBias internally calls
pbaPlotMisclassification, pbaPlotSelection, pbaPlotConfoundingProportions,
and pbaPlotConfoundingRisks.


\code{plotBias}: The plotBias function returns density plots or histograms of the
distribution of bias parameters for a pba object. If print = TRUE, the
function attempts to organize all plots into a single window. Individual 
plots may be accessed from the returned list. plotBias internally calls
pbaPlotMisclassification, pbaPlotSelection, pbaPlotConfoundingProportions,
and pbaPlotConfoundingRisks.


\code{plotBias}: The plotBias function returns density plots or histograms of the
distribution of bias parameters for a pba object. If print = TRUE, the
function attempts to organize all plots into a single window. Individual 
plots may be accessed from the returned list. plotBias internally calls
pbaPlotMisclassification, pbaPlotSelection, pbaPlotConfoundingProportions,
and pbaPlotConfoundingRisks.


\code{plotBias}: The plotBias function returns density plots or histograms of the
distribution of bias parameters for a pba object. If print = TRUE, the
function attempts to organize all plots into a single window. Individual 
plots may be accessed from the returned list. plotBias internally calls
pbaPlotMisclassification, pbaPlotSelection, pbaPlotConfoundingProportions,
and pbaPlotConfoundingRisks.


\code{plotBias}: The plotBias function returns density plots or histograms of the
distribution of bias parameters for a pba object. If print = TRUE, the
function attempts to organize all plots into a single window. Individual 
plots may be accessed from the returned list. plotBias internally calls
pbaPlotMisclassification, pbaPlotSelection, pbaPlotConfoundingProportions,
and pbaPlotConfoundingRisks.


\code{plotBias}: The plotBias function returns density plots or histograms of the
distribution of bias parameters for a pba object. If print = TRUE, the
function attempts to organize all plots into a single window. Individual 
plots may be accessed from the returned list. plotBias internally calls
pbaPlotMisclassification, pbaPlotSelection, pbaPlotConfoundingProportions,
and pbaPlotConfoundingRisks.


\code{plotBias}: The plotBias function returns density plots or histograms of the
distribution of bias parameters for a pba object. If print = TRUE, the
function attempts to organize all plots into a single window. Individual 
plots may be accessed from the returned list. plotBias internally calls
pbaPlotMisclassification, pbaPlotSelection, pbaPlotConfoundingProportions,
and pbaPlotConfoundingRisks.


}
\author{Jeremy Thoms Hetzel \email{jthetzel@gmail.com}}
\arguments{\item{pba}{A pba object.}
\item{density}{Logical; if true, plots smoothed density. If false, plots
histogram.}
\item{sclaes}{Character string passed to ggplot. This parameter controls
whether the x and y axes are free or fixed. Acceptable values are "free",
"fixed", "free_x", or "free_y"}
\item{types}{A character vector specifying the types of biases to plot. If
NULL, the function will plot all defined biases. Acceptable values are
"misclassification", "selection", and "confounding".}
\item{print}{Logical; if true, function prints the plots. If false, the
function only returns the plot objects in a list.}}
\alias{plotBias}

\alias{plotBias}

\alias{plotBias}

\alias{plotBias}

\alias{plotBias}

\alias{plotBias}

\alias{plotBias}

\alias{plotBias}

\alias{plotBias}

