\name{plotEstimates}
\alias{plotEstimates}
\title{plotEstimates}
\usage{plotEstimates(pba, data, density=T, exp=F, adjust=1, binwidth,
    scales="free", variables, print=T)plotEstimates(pba, data, density=T, exp=F, adjust=1, binwidth, scales="free",
    variables, print=T)
plotEstimates(pba, data, density=T, exp=F, adjust=1, binwidth, scales="free",
    variables, print=T)
plotEstimates(pba, data, density=T, exp=F, adjust=1, binwidth, scales="free",
    variables, print=T)
plotEstimates(pba, data, density=T, exp=F, adjust=1, binwidth, scales="free",
    variables, print=T)
plotEstimates(pba, data, density=T, exp=F, adjust=1, binwidth, scales="free",
    variables, print=T)
plotEstimates(pba, data, density=T, exp=F, adjust=1, binwidth, scales="free",
    variables, print=T)
plotEstimates(pba, data, density=T, exp=F, adjust=1, binwidth, scales="free",
    variables, print=T)
plotEstimates(pba, data, density=T, exp=F, adjust=1, binwidth, scales="free",
    variables, print=T)
plotEstimates(pba, data, density=T, exp=F, adjust=1, binwidth, scales="free",
    variables, print=T)
}
\alias{plotEstimates}

\alias{plotEstimates}

\alias{plotEstimates}

\alias{plotEstimates}

\alias{plotEstimates}

\alias{plotEstimates}

\alias{plotEstimates}

\alias{plotEstimates}

\alias{plotEstimates}

