\name{print.pba.variables}
\alias{print.pba.variables}
\title{print.pba.variables}
\usage{print.pba.variables(x, ...)print.pba.variables(x, ...)
print.pba.variables(x, ...)
print.pba.variables(x, ...)
print.pba.variables(x, ...)
print.pba.variables(x, ...)
print.pba.variables(x, ...)
print.pba.variables(x, ...)
print.pba.variables(x, ...)
print.pba.variables(x, ...)
}
\author{Jeremy Thoms Hetzel \email{jthetzel@gmail.com}}
\arguments{\item{x}{An object of class \code{pba.variables}.}

}
\alias{print.pba.variables}

\alias{print.pba.variables}

\alias{print.pba.variables}

\alias{print.pba.variables}

\alias{print.pba.variables}

\alias{print.pba.variables}

\alias{print.pba.variables}

\alias{print.pba.variables}

\alias{print.pba.variables}

