\name{print.pba}
\alias{print.pba}
\title{print.pba}
\usage{print.pba(x, digits=max(3, getOption("digits") - 3), ...)print.pba(x, digits=max(3, getOption("digits") - 3), ...)
print.pba(x, digits=max(3, getOption("digits") - 3), ...)
print.pba(x, digits=max(3, getOption("digits") - 3), ...)
print.pba(x, digits=max(3, getOption("digits") - 3), ...)
print.pba(x, digits=max(3, getOption("digits") - 3), ...)
print.pba(x, digits=max(3, getOption("digits") - 3), ...)
print.pba(x, digits=max(3, getOption("digits") - 3), ...)
print.pba(x, digits=max(3, getOption("digits") - 3), ...)
print.pba(x, digits=max(3, getOption("digits") - 3), ...)
}
\author{Jeremy Thoms Hetzel \email{jthetzel@gmail.com}}
\arguments{\item{x}{An object of class \code{pba}.}
\item{digits}{The number of significant digits to use when printing.}}
\alias{print.pba}

\alias{print.pba}

\alias{print.pba}

\alias{print.pba}

\alias{print.pba}

\alias{print.pba}

\alias{print.pba}

\alias{print.pba}

\alias{print.pba}

