% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roclet-quickcheck.R
\name{quickcheck_roclet}
\alias{quickcheck_roclet}
\title{Roclet: make quickcheck tests.}
\usage{
quickcheck_roclet()
}
\description{
This roclet generates R files in the /tests directory that can make use of 
the \pkg{quickcheck} packages.
}
\seealso{
\code{vignette("quickcheck", package = "roxygen2")}

Other roclets: \code{\link{namespace_roclet}},
  \code{\link{rd_roclet}}, \code{\link{vignette_roclet}}
}

